+++ bin/c-mk-init.js
@@ -1,0 +1,15 @@
+#!/usr/bin/env node
+
+'use strict';
+
+const program = require('commander');
+const { exec } = require('shelljs');
+
+// The basic program, which uses sub-commands.
+program
+    .parse(process.argv);
+
+exec('minikube start --disk-size 64g --extra-config=apiserver.service-node-port-range=80-32767 --kubernetes-version v1.9.4 --memory=4096 --vm-driver=vmware');
+exec('minikube ssh -- docker run -it --rm --privileged --pid=host alpine nsenter -t 1 -m -u -n -i date -u $(date -u +%m%d%H%M%Y)');
+exec('helm init --kube-context minikube');
+exec('helm repo update stable');